DB 문제 수정 진행중 ....

+++ src/main/java/com/teraenergy/illegalparking/model/entity/lawdong/repository/LawDongRepository.java
@@ -10,6 +10,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * Project : illegalParking
  * Description :
  */
-public interface LawDongRepository extends JpaRepository<LawDong, Double> {
+public interface LawDongRepository extends JpaRepository<LawDong, Integer> {
 
 }
